Checking if keyword is a MarkDecorator before set
In NodeKeywords, setitem might be called multiple times for the same
key. For example, pytest.mark.foo('foo') will call setitem first with a
MarkDecorator value and then with a True value. The MarkDecorator value is
therefore overwritten and the marked keyword is never matched. This change
only sets the keyword when it's not already a MarkDecorator instance.