Update README.rst and website sample code with `import pytest`
I am a new user to both Python and pytest. I wrote a basic Python program and refactored functions into a separate file for testing. 
I was only able to execute the tests with `pytest` after I added `import pytest` to the top of the file.
Note importantly that the pytest website should be updated accordingly particularly for beginners, as I discovered this issue when I was following the "Our first test run" example in the [getting started section of the pytest website](http://doc.pytest.org/en/latest/getting-started.html). Another example further down that page in the section "Asserting that a certain exception is raised" actually included `import pytest` which prompted me to added it to my code and I was then able to start using pytest.