Add --pep420-mode to deal with PEP 420
In this mode, the upwards file system traversal when deducing the
test module name is not based on the presence of __init__.py, but
instead only stops if it hits a directory in sys.path.

Users that have PEP 420 namespace packages are supposed to use
this mode, and at the same time make sure they configure their
PYTHONPATH property (typically add the project root directory).