Make the "examples" section more prominent in the docs
I spent some time today figuring out why PR #2881 was not showing up
on doc/parametrize... then after some digging even on readthedocs
wondering if the last documentation build had failed, I realized the
docs I was looking for was in doc/example/parametrize instead.

The section that mentions this is very easy to miss, this makes it a
full fledged title and easier to find.