Revert "Let context-managers for raises and warns handle unknown keyword arguments"
This reverts commit dfe54cd82f55f17f3c9e6e078325f306a046b93b.

The idea in the commit was to simplify the code by removing the check
and instead letting it TypeError which has the same effect.

However this type error is caught by mypy, and rather than ignoring the
error we think it's better and clearer to go back to the previous
explicit check.