MarkGenerator: Report unknown mark warning at caller level
Currently, `PytestUnknownMarkWarning` is reported at the level of
`MarkGenerator`, which is unhelpful to callers as the warning is shown as being
"caused" by pytest. This commit therefore increases the stack level passed to
the warning call to 2, which will make the warning log as being "caused" by the
caller, which is much more helpful.

Resolves #5928.