conftest: allow skipping of test re-ordering
I've wanted this when checking why a certain test is failing when only
being run directly, and thought it would be due to the re-ordering.
Therefore it is useful to being able to skip it via an env.

Adds it to a Travis job, so that we are running tests in two different
orders at least.