Prepare for adding a py.typed file by moving pytest.py to a package
While this commit does not do it yet, we want to eventually publish
pytest's type annotations. According to PEP561, this is done by adding a
`py.typed` file to each top-level package, `pytest` and `_pytest` in
pytest's case. However, the `pytest` package is currently just a single
module/file, so it's not possible to added to file to it.

Change `pytest` to a package (i.e. a directory with `__init__.py`), then
adding a `py.typed` file to it is possible.

The reason for doing it already is that it makes it easy to ask people
to test the types just by touching a couple of files.