pytest_runtestloop, use name removed in 5939b336c
What's needed to fix https://github.com/pytest-dev/pytest/issues/3088
seems to be checking for `rep.when == "setup"`.

This then needs adjustments though (to code done in
https://github.com/pytest-dev/pytest/pull/3110) to make this work with
xdist.

It appears to make sense to split it into hooks that are run per test
(for eventually writing if past edge), and one that gets run at the end
really (not requiring to check for "last item" then).