Remove dynamic setup of `pytest.collect` fake module
Make it a real module instead.

This was initially added in 9b58d6eaca
(https://github.com/pytest-dev/pytest/pull/2315), when removing support
for `pytest_namespace`.

Not sure if it ever worked "properly", but currently
`import pytest.collect` and `from pytest.collect import File` do not
work (but with this patch).

This is only relevant for (backward) compatibility, but the more it
appears to make sense to move it into a lazily loaded, separate file.